HSEARCH-4708 Refactor HibernateOrmPathDefinitionProvider to allow for resolution of extractors
