+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/model/path/spi/PojoModelPathWalker.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.model.path.binding.impl;
+package org.hibernate.search.mapper.pojo.model.path.spi;
 
 import org.hibernate.search.mapper.pojo.model.path.PojoModelPathPropertyNode;
 import org.hibernate.search.mapper.pojo.model.path.PojoModelPathValueNode;
